+++ src/onegov.swissvotes/onegov/swissvotes/upgrade.py
@@ -3,6 +3,9 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 
 """
 from onegov.core.upgrade import upgrade_task
+from onegov.swissvotes.models import SwissVote
+from sqlalchemy import Column
+from sqlalchemy.dialects.postgresql import TSVECTOR
 
 
 @upgrade_task('Rename recommendation columns')
@@ -22,3 +25,15 @@ def rename_recommendation_columns(context):
             context.operations.alter_column(
                 'swissvotes', old, new_column_name=new
             )
+
+
+@upgrade_task('Add tsvector columns')
+def add_tsvector_columns(context):
+    for column in ('searchable_text_de_CH', 'searchable_text_fr_CH', 'french'):
+        if not context.has_column('swissvotes', column):
+            context.operations.add_column(
+                'swissvotes', Column(column, TSVECTOR())
+            )
+
+    for vote in context.app.session().query(SwissVote):
+        vote.vectorize_files()